requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
python-dotenv==1.0.0

# PDF Processing
PyMuPDF==1.23.8
unstructured==0.11.6
unstructured[pdf]==0.11.6
pillow==10.1.0

# LangChain and AI
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.4
openai==1.3.7
tiktoken==0.5.2

# Vector Database
chromadb==0.4.18
sentence-transformers==2.2.2

# Data Processing
pandas==2.1.4
numpy==1.24.3
markdown==3.5.1

# Utilities
pydantic==2.5.0
httpx==0.25.2
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
